Fixed sentence structure
Here the documentation states copilot will ignore files then in the next sentence the documentation states copilot will not ignore file and will be able to see and update them. Ive simply removed the “not” and switched the “and” to an “or” so the documentation makes sense and doesn’t contradict itself.